feat: create local user service
I created a local user service that will be used to authenticate users
using secret passwords and two-factor authentication using TOTP
passwords. I configured the users service to generate the Rust source
code from the Protocol Buffer specification and implement and expose the
gRPC API.

I created a Cargo.toml file in the root directory of the repository. I
set up this Cargo.toml to act as a workspace for the project.