feat: implement the users service with a CLI
I added the clap crate to the local users service to implement support
for passing command line arguments to set the TCP/IP port that the gRPC
server listens on, or to set the port using the APP_PORT environment
variable. This will allow the service to be run on a custom port for
development and testing, or to be used with a Dapr sidecar, which is the
target for the service.

I extracted the gRPC service code from the main.rs file into a new
module. I also added support for returning custom exit codes from the
local users service in the case of errors.